Fixed name clash due to auto-generation process.
