--- groucho-lsh-rest/src/main/java/it/cnr/iasi/saks/groucho/App.java
@@ -1,13 +1,0 @@
-package it.cnr.iasi.saks.groucho;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
